requirements.txt
# requirements.txt - Dependencias actualizadas con autenticación JWT

# FastAPI y dependencies principales
fastapi==0.117.1
uvicorn==0.37.0
python-multipart==0.0.20
python-dotenv==1.1.1
Jinja2==3.1.6
starlette==0.48.0

# Base de datos
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10

# Autenticación JWT (NUEVAS)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Sesiones
itsdangerous==2.2.0

# Validación de datos
pydantic==2.11.9
pydantic_core==2.33.2
email-validator==2.3.0

# Utilidades
PyYAML==6.0.3
certifi==2025.8.3
idna==3.10
dnspython==2.8.0

# CLI y herramientas de desarrollo
fastapi-cli==0.0.13
click==8.3.0
rich==14.1.0
typer==0.19.2
shellingham==1.5.4

# Server y networking
httptools==0.6.4
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
watchfiles==1.1.0
websockets==15.0.1

# Otras dependencias existentes
annotated-types==0.7.0
anyio==4.11.0
colorama==0.4.6
greenlet==3.2.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
Pygments==2.19.2
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.39.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0